requirements.txt
flask_sqlalchemy
flask_script
flask_migrate
flask_bootstrap